android: moved android solution to gradle base

+++ src/graphics/font.cpp
@@ -258,10 +258,10 @@ void add_symbols_to_font_packer(imagepak_handle font_pack, pcstr symbols_font, f
         charset.AddCodepoint(codepoint);
 
         uint8_t colors[4] = { 
-            (fconfig.color >> 16) & 0xff,  // Red
-            (fconfig.color >> 8) & 0xff,   // Green
-            (fconfig.color >> 0) & 0xff,   // Blue
-            (fconfig.color >> 24) & 0xff   // Alpha
+            uint8_t((fconfig.color >> 16) & 0xff),  // Red
+            uint8_t((fconfig.color >> 8) & 0xff),   // Green
+            uint8_t((fconfig.color >> 0) & 0xff),   // Blue
+            uint8_t((fconfig.color >> 24) & 0xff)   // Alpha
         };
         // padding=0, fit=true to create bitmaps with exact glyph dimensions (no extra space)
         DynamicFont::Atlas atlas(symbols_font, fconfig.size, charset, DynamicFont::RenderMode::COLOR, 0, true, colors, fconfig.bold, fconfig.shadow_offset);
